Introduccion al sistema de Peleas, mejoras en los menus

+++ Estructura/Personajes.cs
@@ -6,14 +6,14 @@ namespace CrearPersonajes
         private int fuerza;
         private int nivel;
         private int armadura;
-        private int salud;
+        private float salud;
 
         public int Velocidad { get => velocidad; set => velocidad = value; }
         public int Destreza { get => destreza; set => destreza = value; }
         public int Fuerza { get => fuerza; set => fuerza = value; }
         public int Nivel { get => nivel; set => nivel = value; }
         public int Armadura { get => armadura; set => armadura = value; }
-        public int Salud { get => salud; set => salud = value; }
+        public float Salud { get => salud; set => salud = value; }
     }
     public class Caracteristicas{
         private string tipo;
